기본 repository 작성
