+++ src/main/java/ving/vingterview/domain/board/Board.java
@@ -1,15 +1,18 @@
-package ving.vingterview.domain;
+package ving.vingterview.domain.board;
 
 import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
+import ving.vingterview.domain.EntityDate;
+import ving.vingterview.domain.member.Member;
+import ving.vingterview.domain.question.Question;
 
 @Entity
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class Board extends EntityDate{
+public class Board extends EntityDate {
 
     @Id
     @GeneratedValue
