+++ src/main/java/ving/vingterview/domain/board/BoardMemberLike.java
@@ -1,31 +1,32 @@
-package ving.vingterview.domain;
+package ving.vingterview.domain.board;
 
 import jakarta.persistence.*;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
+import ving.vingterview.domain.member.Member;
 
 @Entity
 @Getter
 @NoArgsConstructor
-public class QuestionMemberScrap {
+public class BoardMemberLike {
 
     @Id
     @GeneratedValue
-    @Column(name = "question_member_id")
+    @Column(name = "board_member_id")
     private Long id;
 
     @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "question_id")
-    private Question question;
+    @JoinColumn(name = "board_id")
+    private Board board;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "member_id")
     private Member member;
 
     @Builder
-    public QuestionMemberScrap(Question question, Member member) {
-        this.question = question;
+    public BoardMemberLike(Board board, Member member) {
+        this.board = board;
         this.member = member;
     }
 }
