+++ src/main/java/ving/vingterview/domain/comment/Comment.java
@@ -1,18 +1,22 @@
-package ving.vingterview.domain;
+package ving.vingterview.domain.comment;
 
 import jakarta.persistence.*;
+import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
+import ving.vingterview.domain.EntityDate;
+import ving.vingterview.domain.member.Member;
+import ving.vingterview.domain.board.Board;
 
 @Entity
 @Getter
-@NoArgsConstructor
-public class BoardMemberLike {
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class Comment  extends EntityDate {
 
     @Id
     @GeneratedValue
-    @Column(name = "board_member_id")
+    @Column(name = "comment_id")
     private Long id;
 
     @ManyToOne(fetch = FetchType.LAZY)
@@ -22,10 +26,13 @@ public class BoardMemberLike {
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "member_id")
     private Member member;
+    private String content;
+
 
     @Builder
-    public BoardMemberLike(Board board, Member member) {
+    public Comment(Board board, Member member, String content) {
         this.board = board;
         this.member = member;
+        this.content = content;
     }
 }
