+++ src/main/java/ving/vingterview/domain/member/Member.java
@@ -1,4 +1,4 @@
-package ving.vingterview.domain;
+package ving.vingterview.domain.member;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
@@ -8,11 +8,12 @@ import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
+import ving.vingterview.domain.EntityDate;
 
 @Entity
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class Member  extends EntityDate{
+public class Member  extends EntityDate {
 
     @Id
     @GeneratedValue
