+++ src/main/java/ving/vingterview/domain/question/Question.java
@@ -1,15 +1,17 @@
-package ving.vingterview.domain;
+package ving.vingterview.domain.question;
 
 import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
+import ving.vingterview.domain.EntityDate;
+import ving.vingterview.domain.member.Member;
 
 @Entity
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class Question extends EntityDate{
+public class Question extends EntityDate {
 
     @Id
     @GeneratedValue
