+++ src/main/java/ving/vingterview/domain/tag/TagMember.java
@@ -1,35 +1,33 @@
-package ving.vingterview.domain;
+package ving.vingterview.domain.tag;
 
 import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
+import ving.vingterview.domain.member.Member;
 
 @Entity
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class Comment  extends EntityDate{
+public class TagMember {
 
     @Id
     @GeneratedValue
-    @Column(name = "comment_id")
+    @Column(name = "tag_member_id")
     private Long id;
 
     @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "board_id")
-    private Board board;
+    @JoinColumn(name = "tag_id")
+    private Tag tag;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "member_id")
     private Member member;
-    private String content;
-
 
     @Builder
-    public Comment(Board board, Member member, String content) {
-        this.board = board;
+    public TagMember(Tag tag, Member member) {
+        this.tag = tag;
         this.member = member;
-        this.content = content;
     }
 }
