+++ src/main/java/ving/vingterview/domain/tag/TagType.java
@@ -1,4 +1,4 @@
-package ving.vingterview.domain;
+package ving.vingterview.domain.tag;
 
 public enum TagType {
     JOB, INTERVIEW, ENTERPRISE, CLASSIFICATION
