+++ src/main/java/ving/vingterview/repository/CommentMemberLikeRepository.java
@@ -1,0 +1,7 @@
+package ving.vingterview.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import ving.vingterview.domain.comment.CommentMemberLike;
+
+public interface CommentMemberLikeRepository extends JpaRepository<CommentMemberLike,Long> {
+}
