+++ src/main/java/ving/vingterview/repository/QuestionMemberScrapRepository.java
@@ -1,0 +1,7 @@
+package ving.vingterview.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import ving.vingterview.domain.question.QuestionMemberScrap;
+
+public interface QuestionMemberScrapRepository extends JpaRepository<QuestionMemberScrap,Long> {
+}
