+++ src/main/java/ving/vingterview/repository/QuestionRepository.java
@@ -1,0 +1,7 @@
+package ving.vingterview.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import ving.vingterview.domain.question.Question;
+
+public interface QuestionRepository extends JpaRepository<Question,Long> {
+}
