+++ src/main/java/ving/vingterview/repository/TagQuestionRepository.java
@@ -1,0 +1,7 @@
+package ving.vingterview.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import ving.vingterview.domain.tag.TagQuestion;
+
+public interface TagQuestionRepository extends JpaRepository<TagQuestion,Long> {
+}
