+++ src/main/java/ving/vingterview/repository/TagRepository.java
@@ -1,0 +1,7 @@
+package ving.vingterview.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import ving.vingterview.domain.tag.Tag;
+
+public interface TagRepository extends JpaRepository<Tag,Long> {
+}
